Bug 1697429 - Handle crash with the old way if the fork server is prefed out. r=jld
Handle crash with the old way if the fork server is prefed out, and waitpid() for all available stat changes of children processes in the forkserver.

Differential Revision: https://phabricator.services.mozilla.com/D110507

UltraBlame original commit: 88ae7d2397ebcb0149ca014e14a75766b35457c3

+++ ipc/chromium/src/base/process_util_posix.cc
@@ -34,6 +34,10 @@
 #include "prenv.h"
 #include "prmem.h"
 
+#ifdef MOZ_ENABLE_FORKSERVER
+#  include "ForkServiceChild.h"
+#endif
+
 const int kMicrosecondsPerSecond = 1000000;
 
 namespace base {
@@ -208,20 +212,22 @@ void CloseSuperfluousFds(void* aCtx, bool (*aShouldPreserve)(void*, int)) {
 
 bool DidProcessCrash(bool* child_exited, ProcessHandle handle) {
 #ifdef MOZ_ENABLE_FORKSERVER
-  
-  
-  
-  
-  
-  const int r = kill(handle, 0);
-  if (r < 0 && errno == ESRCH) {
-    if (child_exited) *child_exited = true;
-  } else {
-    if (child_exited) *child_exited = false;
-  }
+  if (mozilla::ipc::ForkServiceChild::Get()) {
+    
+    
+    
+    
+    
+    const int r = kill(handle, 0);
+    if (r < 0 && errno == ESRCH) {
+      if (child_exited) *child_exited = true;
+    } else {
+      if (child_exited) *child_exited = false;
+    }
 
-  return false;
-#else
+    return false;
+  }
+#endif
   int status;
   const int result = HANDLE_EINTR(waitpid(handle, &status, WNOHANG));
   if (result == -1) {
@@ -263,7 +269,6 @@ bool DidProcessCrash(bool* child_exited, ProcessHandle handle) {
   if (WIFEXITED(status)) return WEXITSTATUS(status) != 0;
 
   return false;
-#endif  
 }
 
 void FreeEnvVarsArray::operator()(char** array) {
